change PlantUML example package from `plantuml` to `shopping`
This is a general example with shopping domain that we can also use for other examples like the upcoming modules API. Having too many similar examples also decreases understandability, so one example to demonstrate modularization via PlantUML and via the new modules API seems better.
