--- archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/plantuml/address/Address.java
@@ -1,7 +1,0 @@
-package com.tngtech.archunit.example.plantuml.address;
-
-import com.tngtech.archunit.example.plantuml.catalog.ProductCatalog;
-
-public class Address {
-    private ProductCatalog productCatalog;
-}
