--- archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/plantuml/xml/types/XmlTypes.java
@@ -1,4 +1,0 @@
-package com.tngtech.archunit.example.plantuml.xml.types;
-
-public class XmlTypes {
-}
