--- archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/plantuml/xml/utils/XmlUtils.java
@@ -1,4 +1,0 @@
-package com.tngtech.archunit.example.plantuml.xml.utils;
-
-public class XmlUtils {
-}
