+++ archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/shopping/address/Address.java
@@ -1,0 +1,7 @@
+package com.tngtech.archunit.example.shopping.address;
+
+import com.tngtech.archunit.example.shopping.catalog.ProductCatalog;
+
+public class Address {
+    private ProductCatalog productCatalog;
+}
