+++ archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/shopping/xml/processor/XmlProcessor.java
@@ -1,0 +1,4 @@
+package com.tngtech.archunit.example.shopping.xml.processor;
+
+public class XmlProcessor {
+}
