+++ archunit-example/example-plain/src/main/java/com/tngtech/archunit/example/shopping/xml/types/XmlTypes.java
@@ -1,0 +1,4 @@
+package com.tngtech.archunit.example.shopping.xml.types;
+
+public class XmlTypes {
+}
