+++ archunit-integration-test/src/test/java/com/tngtech/archunit/integration/ExamplesIntegrationTest.java
@@ -132,12 +132,12 @@ import com.tngtech.archunit.example.onionarchitecture.domain.model.ShoppingCart;
 import com.tngtech.archunit.example.onionarchitecture.domain.service.OrderQuantity;
 import com.tngtech.archunit.example.onionarchitecture.domain.service.ProductName;
 import com.tngtech.archunit.example.onionarchitecture.domain.service.ShoppingService;
-import com.tngtech.archunit.example.plantuml.address.Address;
-import com.tngtech.archunit.example.plantuml.catalog.ProductCatalog;
-import com.tngtech.archunit.example.plantuml.customer.Customer;
-import com.tngtech.archunit.example.plantuml.importer.ProductImport;
-import com.tngtech.archunit.example.plantuml.order.Order;
-import com.tngtech.archunit.example.plantuml.product.Product;
+import com.tngtech.archunit.example.shopping.address.Address;
+import com.tngtech.archunit.example.shopping.catalog.ProductCatalog;
+import com.tngtech.archunit.example.shopping.customer.Customer;
+import com.tngtech.archunit.example.shopping.importer.ProductImport;
+import com.tngtech.archunit.example.shopping.order.Order;
+import com.tngtech.archunit.example.shopping.product.Product;
 import com.tngtech.archunit.exampletest.ControllerRulesTest;
 import com.tngtech.archunit.exampletest.SecurityTest;
 import com.tngtech.archunit.testutil.TransientCopyRule;
